fix: better error message for non-existent directory

+++ scripts/git/git_ignore_all.sh
@@ -22,19 +22,24 @@ readonly force="${force:-false}"
 
 readonly base_dir="${1:-$PWD}"
 
+if [ ! -d "${base_dir}" ]; then
+  printf "The directory '%s' does not exist.\n" "${base_dir}" >&2
+  exit 2
+fi
+
 (
   cd "${base_dir}"
 
   if [ "$(git rev-parse --is-inside-work-tree 2>/dev/null)" != 'true' ]; then
     echo "'${base_dir}' is not a git repository" >&2
-    exit 2
+    exit 3
   fi
 
   if [ "${force}" = 'true' ]; then
     rm -f .gitignore
   elif [ -f '.gitignore' ]; then
     printf "'%s' already exists.\n" "${base_dir}/.gitignore" >&2
-    exit 3
+    exit 4
   fi
 
   cat <<'EOF' >.gitignore
